feat(application): migrate sync operations to shared application APIs
Move sync_pull, sync_push, and sync_init from domain/workflows into
application/use_cases with typed Request/Response/Deps/Error patterns.
The CLI adapter retains infrastructure-specific post-processing
(index rebuild, sync state reset, temp dir cleanup).
